Implement filtering by column
